sign in with account

+++ src/main/kotlin/com/mobile/e2m/account/navigation/AccountGraph.kt
@@ -19,7 +19,7 @@ import com.mobile.e2m.core.ui.navigation.route.AppNavigationRoute
 
 fun NavGraphBuilder.accountNavGraph(
     navController: NavHostController,
-    onAccount: () -> Unit,
+    onAccount: (Int) -> Unit,
 ) {
     beginDestination(
         goToStarted = {
